refactor : useFirebaseInit 훅으로 분리
