+++ hooks/FCM/index.ts
@@ -1,0 +1,2 @@
+export { default as useFCMPushAlarm } from './useFCMPushAlarm';
+export { default as useFirebaseInit } from './useFirebaseInit';
