+++ package.json
@@ -47,7 +47,7 @@
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev": "next dev -p 4002",
+    "dev": "next dev -p 4003",
     "lint": "next lint",
     "start": "next start"
   },
